Ran Eclipse cleanup on fmi code.

+++ org/ptolemy/fmi/FMILibrary.java
@@ -24,13 +24,13 @@
    PT_COPYRIGHT_VERSION_2
    COPYRIGHTENDKEY
 
-*/
+ */
 
 package org.ptolemy.fmi;
+
 import com.sun.jna.Callback;
 import com.sun.jna.Library;
 import com.sun.jna.Pointer;
-import com.sun.jna.TypeMapper;
 
 /**
  * A Java Native Acess (JNA) Wrapper for a Functional Mock-up Unit shared library.
@@ -53,32 +53,32 @@ public interface FMILibrary extends Library {
      */
     public static interface FMIStatus {
         /** The operation completed successfully. */
-        public static final int fmiOK = (int)0;
+        public static final int fmiOK = 0;
 
         /** The operation had a problem, but the computation may
          *  continue.  The logger was called with further information.
          */
-        public static final int fmiWarning = (int)1;
+        public static final int fmiWarning = 1;
 
         /** Discard the current data and try again with a smaller
          *  step size.  Only certain methods can return this
          *  value, see the FMI documentation for details.
          */
-        public static final int fmiDiscard = (int)2;
+        public static final int fmiDiscard = 2;
 
         /** The operation had a problem and fmiFreeModelInstance()
          *  must be called.
          */
-        public static final int fmiError = (int)3;
+        public static final int fmiError = 3;
 
         /** All model instances are corrupted.  No fmi*() methods
          *  may be called.
          */
-        public static final int fmiFatal = (int)4;
-        
+        public static final int fmiFatal = 4;
+
         /** The slave is still executing, but has returned.
-         */   
-        public static final int fmiPending = (int)5;
+         */
+        public static final int fmiPending = 5;
     };
 
     /**
@@ -89,21 +89,21 @@ public interface FMILibrary extends Library {
          * call fmiGetStatus(..., fmiDoStepStatus) to determine if the
          * slave has finished. 
          */
-        public static final int fmiDoStepStatus = (int)0;
+        public static final int fmiDoStepStatus = 0;
 
         /** If fmiDoStep() returns fmiPending, then 
          *  fmiGetStringStatus(..., fmiPendingStatus,...) can be called
          *  to get the status of the asynchronously executing
          *  fmiDoStep().
          */
-        public static final int fmiPendingStatus = (int)1;
+        public static final int fmiPendingStatus = 1;
 
         /** If fmiDoStep() returnd fmiDiscard, then calling
          *  fmiGetRealStatus(..., fmiLastSuccessfulTime,...)
          *  will return the time the last communication step
          *  was successfully computed.
          */
-        public static final int fmiLastSuccessfulTime = (int)2;
+        public static final int fmiLastSuccessfulTime = 2;
     };
 
     /** A callback for the fmiCallbackLogger() function.
@@ -114,7 +114,7 @@ public interface FMILibrary extends Library {
      *
      *  <p>For details about how Callbacks work in JNA, see
      *  <a href="http://twall.github.com/jna/3.4.0/javadoc/overview-summary.html#callbacks">http://twall.github.com/jna/3.4.0/javadoc/overview-summary.html#callbacks</a>.
-     */   
+     */
     public interface FMICallbackLogger extends Callback {
         /** Invoke the fmiCallbackLogger() function.
          *  @param fmiComponent The fmiComponent
@@ -125,12 +125,13 @@ public interface FMILibrary extends Library {
          *  @param message The message in printf format
          */
         // FIXME: Handle vargargs
-        void apply(Pointer fmiComponent, String instanceName, int status, String category, String message/*, Pointer ... parameters*/);
+        void apply(Pointer fmiComponent, String instanceName, int status,
+                String category, String message/* , Pointer ... parameters */);
     };
 
     /** A callback for the fmiCallbackAllocateMemory() function.
      *  See the documentation for FMICallbackLogger above for details.
-     */   
+     */
     public interface FMICallbackAllocateMemory extends Callback {
         /** Invoke the fmiCallbackAllocateMemory() function.
          *  Each byte of the allocated memory should be set to 0.
@@ -143,14 +144,14 @@ public interface FMILibrary extends Library {
 
     /** A callback for the fmiCallbackFreeMemory() function.
      *  See the documentation for FMICallbackLogger above for details.
-     */   
+     */
     public interface FMICallbackFreeMemory extends Callback {
         /** Free the object.
          *  In Java, the Memory.finalize() method will free
          *  up memory for you, so this method is a no-op.
          *  @param object The object to be freed.  If the object
          *  parameter is null, then take no action.
-         */   
+         */
         void apply(Pointer object);
     };
 
@@ -167,5 +168,5 @@ public interface FMILibrary extends Library {
         void apply(Pointer fmiComponent, int status);
     };
 
-    //public TypeMapper TYPE_MAPPER = FMITypeMapper.FMITYPEMAPPER;
+    // public TypeMapper TYPE_MAPPER = FMITypeMapper.FMITYPEMAPPER;
 }
